--- public/global.css
@@ -1,2 +1,0 @@
-
-
