+++ src/lib/components/Accuracy.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { IWord } from '../types/types';
+  import type { IWord } from '$lib/types/types';
 
   export let errors: string[];
   export let currentChars: string[];
