+++ src/lib/components/Caret.svelte
