+++ src/lib/components/Timer.svelte
