+++ src/lib/utils/generate.ts
@@ -1,0 +1,229 @@
+import type { IChar, IWord } from '$lib/types/types';
+
+export function generateWords(count: number) {
+	let wordsArr: IWord[] = [];
+	let charId = -1;
+	for (let id = 0; id < count; id++) {
+		const lastWord: boolean = id === count - 1;
+		let word = listOfWords[Math.floor(Math.random() * listOfWords.length)];
+
+		// if not the last word add a space
+		if (!lastWord) {
+			word = `${word} `;
+		}
+		// split word into chars
+		const chars = word.split('');
+
+		let characters = [];
+		chars.forEach((char) => {
+			let id = (charId += 1);
+			characters.push({ char, id });
+		});
+
+		wordsArr[id] = { word, characters, id };
+	}
+	return wordsArr;
+}
+
+const listOfWords = [
+	'the',
+	'be',
+	'of',
+	'and',
+	'a',
+	'to',
+	'in',
+	'he',
+	'have',
+	'it',
+	'that',
+	'for',
+	'they',
+	'I',
+	'with',
+	'as',
+	'not',
+	'on',
+	'she',
+	'at',
+	'by',
+	'this',
+	'we',
+	'you',
+	'do',
+	'but',
+	'from',
+	'or',
+	'which',
+	'one',
+	'would',
+	'all',
+	'will',
+	'there',
+	'say',
+	'who',
+	'make',
+	'when',
+	'can',
+	'more',
+	'if',
+	'no',
+	'man',
+	'out',
+	'other',
+	'so',
+	'what',
+	'time',
+	'up',
+	'go',
+	'about',
+	'than',
+	'into',
+	'could',
+	'state',
+	'only',
+	'new',
+	'year',
+	'some',
+	'take',
+	'come',
+	'these',
+	'know',
+	'see',
+	'use',
+	'get',
+	'like',
+	'then',
+	'first',
+	'any',
+	'work',
+	'now',
+	'may',
+	'such',
+	'give',
+	'over',
+	'think',
+	'most',
+	'even',
+	'find',
+	'day',
+	'also',
+	'after',
+	'way',
+	'many',
+	'must',
+	'look',
+	'before',
+	'great',
+	'back',
+	'through',
+	'long',
+	'where',
+	'much',
+	'should',
+	'well',
+	'people',
+	'down',
+	'own',
+	'just',
+	'because',
+	'good',
+	'each',
+	'those',
+	'feel',
+	'seem',
+	'how',
+	'high',
+	'too',
+	'place',
+	'little',
+	'world',
+	'very',
+	'still',
+	'nation',
+	'hand',
+	'old',
+	'life',
+	'tell',
+	'write',
+	'become',
+	'here',
+	'show',
+	'house',
+	'both',
+	'between',
+	'need',
+	'mean',
+	'call',
+	'develop',
+	'under',
+	'last',
+	'right',
+	'move',
+	'thing',
+	'general',
+	'school',
+	'never',
+	'same',
+	'another',
+	'begin',
+	'while',
+	'number',
+	'part',
+	'turn',
+	'real',
+	'leave',
+	'might',
+	'want',
+	'point',
+	'form',
+	'off',
+	'child',
+	'few',
+	'small',
+	'since',
+	'against',
+	'ask',
+	'late',
+	'home',
+	'interest',
+	'large',
+	'person',
+	'end',
+	'open',
+	'public',
+	'follow',
+	'during',
+	'present',
+	'without',
+	'again',
+	'hold',
+	'govern',
+	'around',
+	'possible',
+	'head',
+	'consider',
+	'word',
+	'program',
+	'problem',
+	'however',
+	'lead',
+	'system',
+	'set',
+	'order',
+	'eye',
+	'plan',
+	'run',
+	'keep',
+	'face',
+	'fact',
+	'group',
+	'play',
+	'stand',
+	'increase',
+	'early',
+	'course',
+	'change',
+	'help',
+	'line'
+];
