--- src/main.ts
@@ -1,8 +1,0 @@
-import App from './App.svelte'
-
-const app = new App({
-  target: document.getElementById('app'),
-  intro: true,
-});
-
-export default app
