--- src/stores.ts
@@ -1,3 +1,0 @@
-import { writable } from 'svelte/store'
-
-export const position = writable(-1)
